tests: Move MiscMessageTest to test_message_fetch.py.
